+++ packages/gestalt/src/Checkbox/InternalCheckbox.js
@@ -11,7 +11,6 @@ import {
 import classnames from 'classnames';
 import styles from './InternalCheckbox.css';
 import Box from '../Box';
-import colors from '../Colors.css';
 import focusStyles from '../Focus.css';
 import Icon from '../Icon';
 import Label from '../Label';
@@ -101,11 +100,11 @@ const InternalCheckboxWithForwardRef: AbstractComponent<Props, HTMLInputElement>
     setHover(!hovered);
   };
 
-  let bgStyle = colors.whiteBg;
+  let bgStyle = styles.enabled;
   if (disabled) {
-    bgStyle = colors.lightGrayBg;
+    bgStyle = styles.disabled;
   } else if (checked || indeterminate) {
-    bgStyle = colors.darkGrayBg;
+    bgStyle = styles.checked;
   }
 
   let borderStyle = styles.border;